Python/Git_Repo_Spam_Aggregator/requirements.txt
PyGithub==1.55
json
python-decouple==3.4